Move repositories in recorder module
