+++ scripts/import/index.js
@@ -8,7 +8,7 @@ import mime from 'mime';
 import { MongoClient } from 'mongodb';
 import nodeFetch from 'node-fetch';
 
-import Git from '../../src/repositories/git/git.js';
+import Git from '../../src/archivist/recorder/repositories/git/git.js';
 import * as renamer from '../utils/renamer/index.js';
 
 import logger from './logger/index.js';
