+++ scripts/rewrite/rewrite-snapshots.js
@@ -4,8 +4,8 @@ import { fileURLToPath } from 'url';
 import config from 'config';
 
 import Recorder from '../../src/archivist/recorder/index.js';
-import Git from '../../src/repositories/git/git.js';
-import GitRepository from '../../src/repositories/git/index.js';
+import Git from '../../src/archivist/recorder/repositories/git/git.js';
+import GitRepository from '../../src/archivist/recorder/repositories/git/index.js';
 import * as renamer from '../utils/renamer/index.js';
 
 import * as initializer from './initializer/index.js';
