+++ src/archivist/recorder/repositories/git/dataMapper.js
