+++ src/archivist/recorder/repositories/mongo/dataMapper.js
