+++ src/archivist/recorder/repositories/mongo/index.js
