+++ src/archivist/recorder/repositories/repositoryInterface.js
